backend/requirements.txt
# FastAPI core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing and ML
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
shap==0.43.0
joblib==1.3.2

# Database and storage
supabase==2.1.1
boto3==1.29.7

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
kaggle==1.5.16

# HTTP client
httpx==0.25.2
